Update README.md
I am using openHABian 3.3.0 and for Auto Discover of my KLF 200 worked as described. However the discovery of the configured actuators and scenes needed to be manually triggered so I am proposing this documentation update. I got unblocked by this https://community.openhab.org/t/velux-klf-200-binding-not-adding-things-to-inbox/120858 forum thread but thought it might be better to directly address this in the documentation.